naskar/requirements.txt
tweepy
Pillow
openai
requests
